cube work

+++ src/app/controller/controller.cc
@@ -2,7 +2,7 @@
 
 #include "../model/model.hpp"
 
-std::vector<GLdouble> &s21::Controller::GetVertices() {
+std::vector<float> &s21::Controller::GetVertices() {
   s21::Model *model = s21::Model::GetModel();
 
   return model->GetGLVertices();
